pyproject.toml
[project]
name = "rag-research-agent"
version = "0.0.1"
description = "A custom RAG research agent graph in LangGraph."
authors = [
    { name = "Zohaib Aftab", email = "" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.1",    
    "msgspec>=0.18.6",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["retrieval_graph", "index_graph", "shared"]
[tool.setuptools.package-dir]
"langgraph.templates.retrieval_graph" = "src/retrieval_graph"
"langgraph.templates.index_graph" = "src/index_graph"
"retrieval_graph" = "src/retrieval_graph"
"index_graph" = "src/index_graph"
"shared" = "src/shared"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
[tool.pytest.ini_options]
pythonpath = [
  "src"
]